Using Struct for PDF Renderer
I want to add the concept of passing options to renderers, which I
think is going to necessitate passing more data around on the call
stack than I really want to keep doing if applied to the PDF renderer
(which is already passing a somewhat sketchy amount of its state
around and keeping some in parent functions where it really doesn't
belong to allow mutations).  This switches from simple render
ofunctions to render objects, one of which is pdf.Renderer.